A single file create_csv is used instead of 2.
